Limits for request parameters, input, textarea and password fields

+++ APIServer/PwdMan/PwdManController.cs
@@ -16,7 +16,6 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Configuration;
 using APIServer.PwdMan.Model;
 
 namespace APIServer.PwdMan
@@ -24,13 +23,10 @@ namespace APIServer.PwdMan
     [ApiController]
     public class PwdManController : ControllerBase
     {
-        public IConfiguration Configuration { get; }
-
         public IPwdManService PwdManService { get; }
 
-        public PwdManController(IConfiguration configuration, IPwdManService pwdManService)
+        public PwdManController(IPwdManService pwdManService)
         {
-            Configuration = configuration;
             PwdManService = pwdManService;
         }
 
@@ -38,6 +34,7 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/resetpwd")]
         public IActionResult RequestResetPwd([FromBody] string email)
         {
+            if (email?.Length > Limits.MAX_EMAIL_ADDRESS) throw new InputValueTooLargeException();
             PwdManService.RequestResetPassword(email);
             return new JsonResult(true);
         }
@@ -46,6 +43,9 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/resetpwd2")]
         public IActionResult RequestReset([FromBody] UserResetPasswordModel resetPasswordModel)
         {
+            if (resetPasswordModel?.Password?.Length > Limits.MAX_PASSWORD) throw new InputValueTooLargeException();
+            if (resetPasswordModel?.Token?.Length > Limits.MAX_RESETPWD_CODE) throw new InputValueTooLargeException();
+            if (resetPasswordModel?.Email?.Length > Limits.MAX_EMAIL_ADDRESS) throw new InputValueTooLargeException();
             PwdManService.ResetPassword(resetPasswordModel);
             return new JsonResult(true);
         }
@@ -54,6 +54,7 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/register")]
         public IActionResult IsRegisterAllowed([FromBody] string email)
         {
+            if (email?.Length > Limits.MAX_EMAIL_ADDRESS) throw new InputValueTooLargeException();
             return new JsonResult(PwdManService.IsRegisterAllowed(email));
         }
 
@@ -61,6 +62,10 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/profile")]
         public IActionResult RegisterUser([FromBody] UserRegistrationModel userRegistration)
         {
+            if (userRegistration?.Email?.Length > Limits.MAX_EMAIL_ADDRESS) throw new InputValueTooLargeException();
+            if (userRegistration?.Username?.Length > Limits.MAX_USERNAME) throw new InputValueTooLargeException();
+            if (userRegistration?.Token?.Length > Limits.MAX_REGISTER_CODE) throw new InputValueTooLargeException();
+            if (userRegistration?.Password?.Length > Limits.MAX_PASSWORD) throw new InputValueTooLargeException();
             PwdManService.RegisterUser(userRegistration);
             return new JsonResult(true);
         }
@@ -76,6 +81,7 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/confirmation")]
         public IActionResult ConfirmRegistration([FromBody] OutstandingRegistrationModel confirmation)
         {
+            if (confirmation?.Email?.Length > Limits.MAX_EMAIL_ADDRESS) throw new InputValueTooLargeException();
             return new JsonResult(PwdManService.ConfirmRegistration(GetToken(), confirmation));
         }
 
@@ -90,6 +96,7 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/user/unlock")]
         public IActionResult UnlockUser([FromBody] string userName)
         {
+            if (userName?.Length > Limits.MAX_USERNAME) throw new InputValueTooLargeException();
             return new JsonResult(PwdManService.UnlockUser(GetToken(), userName));
         }
 
@@ -97,6 +104,7 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/user")]
         public IActionResult DeleteUser([FromBody] string username)
         {
+            if (username?.Length > Limits.MAX_USERNAME) throw new InputValueTooLargeException();
             return new JsonResult(PwdManService.DeleteUser(GetToken(), username));
         }
 
@@ -132,6 +140,8 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/user/role")]
         public IActionResult UpdateUserRole([FromBody] UserUpdateRoleModel userUpdateRoleModel)
         {
+            if (userUpdateRoleModel?.UserName?.Length > Limits.MAX_USERNAME) throw new InputValueTooLargeException();
+            if (userUpdateRoleModel?.RoleName?.Length > Limits.MAX_ROLE_NAME) throw new InputValueTooLargeException();
             return new JsonResult(PwdManService.UpdateUserRole(GetToken(), userUpdateRoleModel));
         }
 
@@ -146,6 +156,8 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/auth")]
         public IActionResult Login([FromBody] AuthenticationModel authentication)
         {
+            if (authentication?.Username?.Length > Limits.MAX_USERNAME) throw new InputValueTooLargeException();
+            if (authentication?.Password?.Length > Limits.MAX_PASSWORD) throw new InputValueTooLargeException();
             var ipAddress = HttpContext.Connection.RemoteIpAddress.ToString();
             return new JsonResult(PwdManService.Authenticate(authentication, ipAddress));
         }
@@ -154,6 +166,7 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/auth2")]
         public IActionResult LoginPass2([FromBody] string totp)
         {
+            if (totp?.Length > Limits.MAX_2FA_CODE) throw new InputValueTooLargeException();
             return new JsonResult(PwdManService.AuthenticateTOTP(GetToken(), totp));
         }
 
@@ -176,6 +189,8 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/userpwd")]
         public IActionResult ChangeUserPassword([FromBody] UserPasswordChangeModel userPasswordChange)
         {
+            if (userPasswordChange?.OldPassword?.Length > Limits.MAX_PASSWORD) throw new InputValueTooLargeException();
+            if (userPasswordChange?.NewPassword?.Length > Limits.MAX_PASSWORD) throw new InputValueTooLargeException();
             PwdManService.ChangeUserPassword(GetToken(), userPasswordChange);
             return new JsonResult(true);
         }
@@ -184,6 +199,21 @@ namespace APIServer.PwdMan
         [Route("api/pwdman/file")]
         public IActionResult SavePasswordFile([FromBody] PasswordFileModel passwordFile)
         {
+            if (passwordFile?.SecretKey?.Length > Limits.MAX_PWDMAN_KEY) throw new InputValueTooLargeException();
+            if (passwordFile?.Passwords != null)
+            {
+                foreach (var pwdItem in passwordFile.Passwords)
+                {
+                    if (pwdItem.Name?.Length > Limits.MAX_PWDMAN_NAME ||
+                        pwdItem.Login?.Length > Limits.MAX_PWDMAN_LOGIN ||
+                        pwdItem.Url?.Length > Limits.MAX_PWDMAN_URL ||
+                        pwdItem.Password?.Length > Limits.MAX_PWDMAN_PASSWORD ||
+                        pwdItem.Description?.Length > Limits.MAX_PWDMAN_DESCRIPTION)
+                    {
+                        throw new InputValueTooLargeException();
+                    }
+                }
+            }
             PwdManService.SavePasswordFile(GetToken(), passwordFile);
             return new JsonResult(true);
         }
